move test support to yet another crate
