+++ Cargo.toml
@@ -9,6 +9,7 @@ members = [
     "linkerd/app/integration",
     "linkerd/app/outbound",
     "linkerd/app/profiling",
+    "linkerd/app/test",
     "linkerd/app",
     "linkerd/cache",
     "linkerd/buffer",
