+++ linkerd/app/test/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "linkerd2-app-integration"
+name = "linkerd2-app-test"
 version = "0.1.0"
 authors = ["Linkerd Developers <cncf-linkerd-dev@lists.cncf.io>"]
 edition = "2018"
