+++ linkerd/app/test/src/data/bar-ns1/ca1-cert.pem
